Edit migrations and seed.rb

+++ db/migrate/20181025191127_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[5.2]
   def change
     create_table :orders do |t|
       t.integer :quantity
-      t.decimal :total_price
+      t.integer :total_price
       t.references :user, foreign_key: true
       t.references :farmer, foreign_key: true
